+++ src/main/java/khanh/book_ecommerce/services/OrderService.java
@@ -5,5 +5,5 @@ import khanh.book_ecommerce.models.OrderDetail;
 import java.util.List;
 
 public interface OrderService {
-    public List<OrderDetail> findOrderDetailsByOrderId(int orderId);
+    public List<OrderDetail> findOrderDetailsByOrderId(long orderId);
 }
